refactor: Moved IngestEventService from subfolder emitters to services.
